Test that the extractor's custom behavior is used
This makes the default mock extractor return an instance of a
custom type defined in the test module, so that a bug where the
code under tests accidentally does the same specific thing that
the tests ask it to do (with the intention of testing more general
behavior) is much less likely to go uncaught.

Previously, it extracted the event arguments as a tuple of them,
which was more likely to match a plausible bug.