Copy related implementation from other project
The added code is from a pull request I opened on
https://github.com/dmvassallo/EmbeddingScratchwork.

The PR, https://github.com/dmvassallo/EmbeddingScratchwork/pull/74,
is subject to change (whether or not it is eventually merged).

The version of my code that the additions in this commit are copied
from should remain available (at least longer) at:

https://github.com/EliahKagan/EmbeddingScratchwork/blob/b40a48a41358778333b06cc8c2914ecc6dce8e1c/tests/_audit.py

The copied code won't satisfy the requirements expressed in the
unit tests, without modifications. In particular, there is no Hook
class (it uses a single global audit hook and global table mapping
events to their listeners. I'm mainly committing this to have this
history of the idea in the history.