Unsuccessful attempt to type hint mocked_subscribe_unsubscribe_cls
I consider this attemp to add static typing to the
mocked_subscribe_unsubscribe_cls fixture to be unsuccessful.

mypy actually accepts this, but pyright does not. (Without this,
mypy rejected the code, while pyright accepted it.) It is also
cumbersome and somewhat hard to reason about.

I think a different design for fixture is in order. I always do the
same kinds of things with this fixture: access the class's method
mocks unbound to do assertions with them, and instaniate the class
to use the indirect Hook instance. A fixture could aggregate those
three things (the mocks and and instance made by calling the class)
and bind some kind of aggregate of them to the test function's
parameter named after the fixture, instead of binding the weird
local class (as is currently done).