Change plan: lock in Hook's sub/unsub, no SyncHook
Instead of having a separate SyncHook class, I'm instead having
Hook lock in its subscribe and unsubscribe methods (not in the
installed auditing event hook, for performance reasons; as
commented, and as sort of documented in the docstring, the code is
written in a way that *should* be safe to not do that, even though
the installed auditing event hook can be called at any time,
including while subscribe or unsubscribe is running).

The factory for the mutex can optionally be passed by the user, so
if locking is unwanted, contextlib.nullcontext can be passed. If
no factory is specified, then locking.Lock will be used.

This starts on the plan, but further changes need to be made. In
particular, this functionality needs to be tested!

FIXMEs on the Hook class are largely either done or made obsolete
now, and removed accordingly, leaving one, in modified form. One
that I removed but could potentially still be relevant has to do
with future directions in handling large numbers of listeners on
the same event, by using a different data structure than a tuple,
once the number gets big. But I don't think it's especially
important that any discussion of this -- beyond covering the
limitations of the current implementation -- be documented.