Have subscribe/unsubscribe method mocks wrap behavior
So they have the side effects of actually subscribing and
unsubscribing.

This also makes it so that using them in ways the base-class
subscribe and unsubscribe cannot be used, besides to call
assertion methdos and other stuff Mock adds with a specific
meaning, causes an error. That makes the tests a big more robust.