Draft of scaling test
This is not finished. It may be raising the event more often than
necessary (both in terms of what kind of state churn I want to test
and in terms of being able to understand the report if an assertion
fails). Also, although it requests the subtests fixture with the
intention of having a subtest that asserts something about timing,
I'm not yet doing that.