Check effect of command-line redirection on event
The builtins.input/result event is prevented by even redirection
set up by a shell, as shown by running scripts/simple.py
interactively and then with redirection. (Like manually changing
what is open on file descriptor 0 as I had done before, this is
also *real* redirection, and a more common way of doing it: the
shell changes what is open on that file descriptor in a forked
process before calling one of the functions in the exec family.)

The code in the input builtin's implenentation (in C) that raises
builtins.input/result seems to only run when readline is being
used, which is when stdin is a tty. (In view of the description in
https://docs.python.org/3/library/functions.html#input of when
the builtin raises the event--that it "Raises an auditing event
builtins.input/result with the result after successfully reading
input"--this looks like a bug in CPython.) Specifically, the call
`PySys_Audit("builtins.input/result", "O", result)` is nested
inside a large block governed by `if (tty)`, which is itself
commented "If we're interactive, use (GNU) readline".

This commit moves the redirect.py script into a newly created
scripts/ directory, and creates the new simple.py script in that
same location. I intend to delete both scripts before merging this
branch to main.