Simplify test_can_listen_to_standard_events_for_input
This won't work on CPython, but it didn't before either. I could
make a more complicated test using pexpect (and wexpect on
Windows) that would do it, but that may be excessive, if the
current CPython behavior is to be considered a bug.

This may still need to be revised, and it definitely does need to
be marked xfail on CPython with a message summarizing or referring
to the issue and mentioning or otherwise making clear that the
test passes on PyPy.