Finish planned expansion of "real" event examples
This includes fixing the factual inacuracies I had inadvertently
introduced before and had added a fixme about (also removed now).

I don't necessarily intend that this be the final form of the
notebooks! But this commit completes the specific plan to correct
and expand the claims made and to fix some bugs in them.