Really make test_subaudit runnable as a script
Even though I think I'm going to remove this feature very soon, I
may as well have it working as well as I had thought I did.

Note that as well as I thought (and now have them running) is not
actually very well, because if an imported module has assert
statements in it, the opportunity for pytest to rewrite them for
proper introspection has passed if they are imported before
`if __name__ == '__main__': ...` runs.