Remove feature to run test modules as scripts
It's not worth bringing back this feature and taking the time to
make sure it really works properly at least until I've determined
which fixtures are going in a shared conftest module.

The reason is that, if some of those fixtures provide assertions
(such as _ReprAsserter) rather than the more common arrangement and
cleanup logic (or other functionality), then simple techiques may
not be adequate. That's because pytest's assert-statement rewriting
cannot rewrite code of modules that have already been imported at
the time pytest starts to run and recognized that it should attempt
to rewrite them.

It's possible to use the subprocess module to work around this, and
there may be other techniques. But I'm leaning toward thinking the
extra complexity is not justified, since I rarely actually run test
modules as scripts.