Remove listeners using tuple (not list) operations
Previously, I was reversing it to a list, calling remove, and
reversing it back to a tuple. Now I reverse it to a tuple and
use (non-reversing) slicing and concatenation on the original,
based on the discovered last index.

This seems to make little or no difference for CPython, but to
slightly improve speed in the churn tests for PyPy. The
improvement is by up to 10% but sometimes less. I have not tested
extensively.