Split dev into test and analyze (fixes pypy3.7 tox)
Some of the linting and type checking tools ("analyzers") depend on
typed-ast on Python 3.7. But typed-ast doesn't install on PyPy 3.7.
(Or I don't know how to make it instal on it, at least on Unix-like
systems.)

Splitting all those tools (even those that never need typed-ast)
out of dev as a new "analyze" group, and renaming the dev group
"test" (since now all the packages in it are for unit testing)
maintains convenience and conceptual clarity, while working around
that problem and enabling tests to run successfully on all the
PyPy versions listed in tox.ini.