Rewrite comment about getting direct str instance
Even without event being a direct instance of str, things should
work on CPython, but I still want to "normalize" the event name in
this way, at least for now.

This is more so the assumptions stated in comments in _hook, and in
the readme, are accurate, i.e., so that if another implementation
has weaker atomicity guarantees than CPython but satisfies the
stated prerequisites, then the code is really free of a race
condition (in the absence of other bugs).

I don't think it's worth writing this all out in the function, and
the behavior may change later anyway.