Bring back @skip_if_unavailable testing notebook
At least for now.

This reverts ec46f7d to bring back that notebook, changes the
notebook name and moves it into a new location in accordance with
the current project structure, and adds an explanation at the top
of the notebook to explain its relationship to other documentation
and code in the project (and to discourage its use when something
else would be more suitable, which is almost always).