Stop listing ipython as a direct dependency
It is still guaranteed to be installed as a dependency of
ipykernel.

I added it in 4de87d4 because poetry was computing a version for it
as an indirect dependency (in poetry.lock) that was not compatible
with Python 3.8. But this was due to a bug in IPython 8.13.0
(https://github.com/ipython/ipython/issues/14053), which has been
fixed, so poetry no longer needs help calculating the dependencies.