Point readme section links to GitHub explictly
So that they do something useful on the PyPI page

This makes them worse when viewed in some ways where they would
work, such as the VS Code preview. In addition, it is less than
ideal for forks. A better solution will be to use a separate readme
file on PyPI, generated automatically from the main readme. But for
now, I think this workaround is better than nothing.