Use faster macOS runner
The churn test is slower with PyPy than CPython (it is written in a
way intended to be as fast as reasonable on CPython), which on the
macOS runnner makes it not complete in the asserted time.

This is an imperfect solution because really the code should be
made faster on PyPy (if necessary, by defining different versions
of the function dependeing on the Python implementation).