Update dependencies (#4)
The most important dependency update here is tornado (which is a
"development" dependency, specifically from the "notebook" group,
due to being an indirect dependency of ipykernel, which is listed
in that group). Version 6.3.1 was vulnerable to CVE-2023-28370 (a
case of CWE-601). Version 6.3.2, which is one of the dependencies
updated here, fixes that.