Drop PyPy 3.7 from tox and CI
PyPy 3.8 and PyPy 3.9 are still supported and are not removed.

This also doesn't do anything with CPython 3.7. The Python Software
Foundation has recently stopped supporting that, and I'll likely
remove it from tox and CI once it becomes inconvenient to maintain
it, but I'm not rushing to do that at this time.