Better way to specify experimental 3.12 on CI
This lists the version as just 3.12, but uses the allow-prereleases
key to let setup-python match that to prereleases including the
current release candidate.

While I'm at it, I've also set continue-on-error for the 3.12
test job, so that it failing won't stop any other jobs from
running, since it's more likely to fail for reasons where I would
still want to see the output of the other jobs in order to
understand if the failure is 3.12-specific.

Both these ways of treating 3.12 specially should be undone
sometime after 3.12.0 is out as a stable release, possibly very
soon afterward.