Test on more Pythons and OSes
With some combinations removed that are not (or not yet) working,
either due to the unavailability of the interpreter (in old
versions), or a failure building binary wheels for dependencies of
`poetry` when installing `poetry`. In both cases, this does not
mean `subaudit` wouldn't work, but it prevents it from being tested
on CI using the specific technique for setting up tools that is
currently being used.

Although the overall effect of this change is to test far more
combinations of operating system, Python implementation, and Python
version, this does also remove from testing some combinations with
PyPy 3.8 that had been working when introduced but have stopped
working. Hopefully this and other old versions can be restored (if
`subaudit` is going to continue trying to support them), but even
if they are not, I don't expect this to be a problem because this
also adds multiple newer versions of PyPy that weren't available
when the workflow and `tox.ini` were written. So PyPy coverage is
reasonably good.

(Two notable limitations here, not related to PyPy, are that this
still does not test any free-threaded interpters such as 3.13t, and
it does not test GraalPy. Both would be valuable to test.)