Run type checkers on multiple Pythons
This runs `mypy` and `pyright` on both CPython and PyPy, and on all
versions being tested, both locally and on CI.

This fails on CI (and probably locally as well) for PyPy 3.7, and
otherwise passes. The reason for the failure on PyPy 3.7 is that
`mypy` and `pyright` both depend on `typed-ast`, which contains C
code that does not build on PyPy 3.7.

(That will be worked around in a separate commit, probably by just
not running them on PyPy 3.7.)