Don't attempt `mypy` and `pyright` on PyPy 3.7
Because they both depend on `typed-ast`, which has to buid C code
to install, and that code has a build error on PyPy 3.7.

Other combinations of implementation and version are unaffected. In
particular, they work fine on CPython 3.7.

Although the change is only in `tox.ini`, this change applies both
to local runs and CI runs.

This also slightly further tweaks the whitespace style in `tox.ini`
for readability.