Let the CI workflows also be triggered manually
This adds the `workflow_dispatch` event trigger to both the `tox`
and `markdown` workflows, so they can be triggered by clicking a
button in the Actions tab (as well as the other events, which
shall remain the most common ways they are run).

This is mainly useful for the `markdown` workflow, but it might
possibly come in handy for the `tox` workflow as well.