Test Python 3.14, and change `macos-13` runner to `macos-15-intel` (#91)
On using `macos-15-intel` instead of `macos-13`, see:
https://github.com/actions/runner-images/issues/13046

This should fix the failures that occurred in #89 and #90. This
intentionally does not also do the updates that those PR should be
able to do once this is merged. This likewise intentionally does
not attempt to to have CI test the free-threaded Python 3.13 or the
free-threaded Python 3.14, even though one or both of those may be
working by now; that can also be put off to its own later PR.

Other changes:

- Disable PyPy 3.9 on Windows, while keeping both earlier and
  later versions (as well as CPython 3.9). PyPy version 3.9
  specifically has started failing when installing `poetry`
  due to an error while building the `zstandard` dependency whose
  message indicates it could be solved by installing `setuptools`,
  but that is already being done. There are various avenues that
  could be explored to get this working, but for now I just
  exclude the job. For details on the error, see:
  https://github.com/EliahKagan/subaudit/actions/runs/20082396227/job/57612320648

- Fix a typo where I had wrongly commented `amd64` as being a
  synonym of `aarch64`, but had meant `arm64`.

- Set `fail-fast: true` explicitly so it's easy to find where to
  change it to `fail-fast: false` when needed for debugging.